requirements.txt
fastapi[standard]
SQLAlchemy
pydantic-settings
SQLAlchemy-Utils
psycopg2-binary
alembic
python-dotenv